requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
python-dotenv
langgraph